+++ assets/app/app/audiodata/uploads/upload.js
@@ -279,15 +279,22 @@ angular.module('a2.audiodata.uploads.upload', [
         });
     };
 
+    $scope.getCountOfUploaded = function() {
+        if (!$scope.uploader || !$scope.uploader.queue || !$scope.uploader.queue.length) return 0;
+
+        $scope.uploaded = $scope.uploader.queue.filter(function(file) {
+            return !!file.isUploaded && !file.isError;
+        }).length
+        return $scope.uploaded
+    };
+
     $scope.clearQueue = function() {
         $scope.uploader.progress = 0;
         $scope.uploaded = 0
-    }
+    };
 
     $scope.uploaded = 0;
-    $scope.uploader.onProgressAll = function() {
-        $scope.uploaded = Math.floor($scope.uploader.progress/100 * $scope.uploader.queue.length);
-    };
+    $scope.uploader.onProgressAll = function() {};
 
 })